+++ CHANGELOG.md
@@ -567,3 +567,4 @@
 * Generate a default wallet for every chain when generating vault
 * Add fallback for recover vault - the default address to be recovered for every chain
 * Updated recover vault to create default wallet address for bitcoin for preexisting vaults
+* Updated test result for bitcoin